@@ -1,6 +1,6 @@
 buildscript {
   ext {
-    supportLibrary = "27.1.1"
+    supportLibrary = '1.0.0-beta01'
     kotlinVersion = '1.2.60'
   }
   repositories {
@@ -26,9 +26,9 @@ task clean(type: Delete) {
 }
 
 ext {
-  compileSdkVersion = 27
+  compileSdkVersion = 28
   minSdkVersion = 16
-  targetSdkVersion = 27
+  targetSdkVersion = 28
   versionCode = 13
   versionName = "1.3.2"
 
