@@ -3,9 +3,9 @@ package com.droibit.android.customtabs.launcher.example
 import android.content.ActivityNotFoundException
 import android.net.Uri
 import android.os.Bundle
-import android.support.customtabs.CustomTabsIntent
-import android.support.v4.content.ContextCompat
-import android.support.v7.app.AppCompatActivity
+import androidx.browser.customtabs.CustomTabsIntent
+import androidx.core.content.ContextCompat
+import androidx.appcompat.app.AppCompatActivity
 import android.view.View
 import android.widget.Toast
 import com.droibit.android.customtabs.launcher.CustomTabsLauncher
